Create tokenizers for each language

+++ src/data/03_create_tokenizers.py
@@ -9,10 +9,12 @@ from tqdm import tqdm
 import click
 import numpy as np
 
+from src.data import LANGUAGES
+
 
 def get_vectors(path):
     vectors = dict()
-    with open(path, 'r') as fhandle:
+    with open(path, 'r', encoding='utf-8') as fhandle:
         for i, line in enumerate(fhandle):
             elements = line.split()
             if len(elements) > 2:
@@ -24,13 +26,15 @@ def get_vectors(path):
                     print("Could not process line {}".format(i))
     return vectors
 
+def create_tokenizer(lang):
+    logging.info(f"Working on {lang}.")
+    train_file = f'./data/processed/europarl/europarl.tokenized.{lang}.split/train/data.txt'
+    output_path = f'./data/processed/europarl-{lang}-tokenizer/'
 
-@click.command()
-@click.option('--train-file', default='./data/processed/europarl/europarl.tokenized.en.split/train/data.txt')
-@click.option('--output-path', default='./data/processed/gpt2-europarl-tokenizer/')
-def main(train_file, output_path):
+    logging.info(f"Loading word vectors.")
+    ft_vectors = get_vectors(f'./data/raw/embeddings/wiki.{lang}.align.vec')
+
     types = defaultdict(int)
-    ft_vectors = get_vectors('./data/raw/embeddings/wiki.en.align.vec')
 
     # We will grab every word that has a vector either in the train or valid
     with open(train_file) as fhandle:
@@ -60,18 +64,25 @@ def main(train_file, output_path):
     vocab = {"<pad>": 0, '<unk>': 1, "</s>": 2}
     for word in sorted(types):
         vocab[word] = len(vocab)
-    with open("./data/processed/europarl.vocab", 'w') as fhandle:
+    with open(f"./data/processed/europarl-{lang}.vocab", 'w') as fhandle:
         for word, idx in vocab.items():
             fhandle.write("{} {}\n".format(word, idx))
 
     tokenizer = TransfoXLTokenizer(
-        vocab_file="./data/processed/europarl.vocab",
+        vocab_file=f"./data/processed/europarl-{lang}.vocab",
         eos_token="</s>"
     )
 
-    shutil.rmtree(output_path)
+    if os.path.exists(output_path) and os.path.isdir(output_path):
+        shutil.rmtree(output_path)
     os.makedirs(output_path)
     tokenizer.save_pretrained(output_path)
+
+
+def main():
+    for lang in LANGUAGES:
+        create_tokenizer(lang)
+
 
 
 if __name__ == '__main__':
